add run and save methods and moved input manager to a Managers script with Task Manager
