+++ utils/histogramming/Managers.py
@@ -1,6 +1,8 @@
 from utils.common.tools import h5py_to_ak, json_to_ak, parquet_to_ak
-from utils.histogramming.objects import WeightSyst
+from utils.histogramming.objects import Observable, _Binning, _Systematic, NTupSyst, TreeSyst, WeightSyst, CrossProduct
+from glob import glob 
 import dask
+import os
 from collections import defaultdict
 from pprint import pprint
 import boost_histogram as bh
@@ -12,16 +14,92 @@ import re
 from utils.common.logger import ColoredLogger
 
 logger = ColoredLogger()
+
+class _InputManager(object):
+    def __init__(self, xps, cfg):
+        read_methods = {
+                        'parquet': 'ak_parquet',
+                        'json': 'ak_json',
+                        'root': 'uproot',
+                        'h5': 'ak_h5'
+                        }
+        self.xps = xps
+        self.indirs = cfg["general"]["indir"]
+        self.from_pyth = cfg["general"]["frompythium"]
+        self.ext = cfg["general"]["informat"]
+        self.sample_sel = cfg["general"]["samplesel"]
+        self.reader = read_methods[self.ext]
+    
+    def required_variables(self):
+        req_vars: List[Observable] = []
+        xp_to_req = defaultdict(list)
+        for xp in self.xps:
+            sample, region, obs, systematic, template = xp
+            required_variables = []
+            obs_vars, _region_sel_vars, sample_sel_vars, syst_vars = [],[],[],[]
+            
+            if obs.builder is None:   obs_vars = [obs]
+            else:
+                obs_vars = [Observable(reqvar, reqvar, obs.binning, obs.dataset) for reqvar in obs.builder.req_vars ]
+
+            region_sel_vars =  [ Observable(reqvar, reqvar, obs.binning, obs.dataset) for reqvar in region.sel.req_vars ] 
+            if self.sample_sel:
+                sample_sel_vars =  [ Observable(reqvar, reqvar, obs.binning, obs.dataset) for reqvar in sample.sel.req_vars ]
+            
+            if isinstance(systematic, WeightSyst):
+                template = getattr(systematic, template)
+                if isinstance(template, Functor):
+                    syst_vars =  [ Observable(reqvar, reqvar, obs.binning, obs.dataset) for reqvar in template.req_vars ] 
+                else:
+                    syst_vars = [Observable(template, template, obs.binning, obs.dataset)]  
+            
+            required_variables.extend(obs_vars+region_sel_vars+sample_sel_vars+syst_vars)
+            
+            xp_to_req[xp] = required_variables
+            
+        return xp_to_req
+
+    def required_paths(self):
+        xp_to_paths = defaultdict(list)
+        for xp in self.xps:
+            sample, region, observable, systematic, template = xp
+            if self.from_pyth:  # Follow pythium naming scheme
+                sample_name = sample.name
+                obs_dataset = observable.dataset
+                paths = [f"{path}/{sample_name}_*_{obs_dataset}.{self.ext}" for path in self.indirs]
+                if template != 'nom':
+                    if isinstance(systematic, NTupSyst):
+                        sys_dirs =    getattr(systematic, "where")
+                        sys_samples = getattr(systematic, template)
+                        if sys_dirs == [None]:
+                            paths = [f"{indir}/{s_samp}_*_{obs_dataset}.{self.ext}" for indir in indirs for s_samp in sys_samples]
+                        else:
+                            paths = [f"{s_dir}/{s_samp}_*_{obs_dataset}.{self.ext}" for s_dir in sys_dirs for s_samp in sys_samples]
+                    elif (isinstance(systematic, TreeSyst)):
+                        syst_dataset = getattr(systematic, template) 
+                        paths = [f"{indir}/{sample_name}_*_{syst_dataset}.{self.ext}" for indir in self.indirs]
+
+                paths = list(set([p for path in paths for p in glob(path) if not os.path.isdir(p) ]))
+                xp_to_paths[xp] = paths
+               
+            
+            else:
+                ## TODO:: Assume user defined (somehow?) #Custom? some supported special types? decoreator for custom?
+                logger.warning("Only outputs from Pythium currently supported")
+                pass
+
+        return xp_to_paths
+
 class _TaskManager(object):
-    def __init__(self, outdir, method):
+    def __init__(self, method, sample_sel):
         readers = {
                     'ak_parquet': parquet_to_ak,
-                    'ak_jsob': json_to_ak,
+                    'ak_json': json_to_ak,
                     'ak_h5': h5py_to_ak,
                     'uproot': NotImplementedError("Unsupported input type")
                     }
         self.reader = readers[method]
-        self.outdir = outdir
+        self.sample_sel = sample_sel
     
 
     @dask.delayed
@@ -79,15 +157,15 @@ class _TaskManager(object):
         region = xp["region"]
         # 1st apply overall cuts
         # TODO:: What about overall cuts to apply to all histos?
-        # then apply sample cuts
-
-        if sample.sel is not None:
+        
+        # 2. Apply sample cuts
+        if sample.sel is not None and self.sample_sel:
             new_data = new_data[sample.sel.evaluate(new_data)]
 
-        # then apply region cuts
+        # 3. Apply region cuts
         new_data = new_data[region.sel.evaluate(new_data)]
         
-        # TODO:: then apply observable cuts
+        # TODO:: 4. Apply observable cuts
         
         return new_data
     
@@ -154,12 +232,14 @@ class _TaskManager(object):
                 xp_to_hists[xp].append(self._make_histogram(var, weights,xp))
 
         
-        jobs = []
-        for xp, data in xp_to_hists.items():    jobs.append(data)
+        jobs, xps = [], []
+        for xp, data in xp_to_hists.items():    
+            jobs.append(data)
+            xps.append(xp)
 
-        dask.visualize(jobs, filename=f'{self.outdir}/task_graph.png')
         
-        return jobs
+        
+        return jobs, xps
 
     @classmethod
     def hist_wanted(cls, sample, region, observable, syst, template, ):
