+++ utils/histogramming/processor.py
@@ -1,27 +1,28 @@
 from pathlib import Path
-from utils.histogramming.TaskManager import _TaskManager
-from utils.histogramming.objects import Observable, _Binning, _Systematic, NTupSyst, TreeSyst, CrossProduct
+from utils.histogramming.Managers import _TaskManager, _InputManager
+from utils.histogramming.objects import Observable, _Binning, _Systematic, NTupSyst, TreeSyst, WeightSyst, CrossProduct
 from utils.common.functor import Functor
-from glob import glob 
 import os
 from collections import defaultdict
 from pprint import pprint
 import dask
 import boost_histogram as bh
+import pickle
 
 class Processor(object):    
     def __init__(self, config):
         self.cfg = config
-
+        self.outdir = Path(self.cfg["general"]["outdir"])
+        os.makedirs(self.outdir, exist_ok=True)
+        self.graph = None
+        self.xps = None
+    
     def get_input_files(self, sample, observable, systematic, template,):
         indirs = self.cfg["general"]["indir"]
         from_pyth = self.cfg["general"]["frompythium"]
         ext = self.cfg["general"]["informat"]
 
     def create(self):
-        #======= Output location of histogram
-        hist_folder = Path(self.cfg["general"]["outdir"])
-        os.makedirs(hist_folder, exist_ok=True)
         #======= Analysis objects from config 
         samples =  self.cfg["samples"]
         regions = self.cfg["regions"]
@@ -30,15 +31,35 @@ class Processor(object):
 
         
         xp_iter = list(self.cross_product(samples, regions, systematics, observables))
-        input_manager =  InputManager(xp_iter, self.cfg)
+        input_manager =  _InputManager(xp_iter, self.cfg)
         xp_to_req_vars = input_manager.required_variables()
         xp_to_paths = input_manager.required_paths()
 
-        task_manager = _TaskManager(hist_folder, input_manager.reader)
-        task_tree= task_manager._build_tree(xp_to_paths, xp_to_req_vars)
-        histograms = dask.compute(*task_tree, scheduler = "synchronous")
+        task_manager = _TaskManager(input_manager.reader, sample_sel = self.cfg["general"]["samplesel"])
+        task_tree, xps = task_manager._build_tree(xp_to_paths, xp_to_req_vars)
+        dask.visualize(task_tree, filename=f'{self.outdir}/task_graph.png')
+        
+        self.graph = task_tree
+        self.xps = xps
 
-    
+    def run(self):
+        histograms = dask.compute(*self.graph, scheduler = "synchronous")
+        dd = defaultdict(dict)
+        for i, xp in enumerate(self.xps):
+            sample, region, obs, syst, template = xp
+            if syst is not None:
+                dd[obs.name].update({f'{sample.name}_{region.name}_{syst.name}_{template}': histograms[i]})
+            else:
+                dd[obs.name].update({f'{sample.name}_{region.name}_{template}': histograms[i]})
+        return dd
+
+    def save(self, hists_dict):
+        for obs in list(hists_dict.keys()):
+            fname = f'{obs}.pkl'
+            with open(f"{self.outdir}/{fname}", "wb") as f:
+                pickle.dump(hists_dict[obs], f)
+
+
     def cross_product(self, samples, regions, systematics, observables):
         for sample in samples:
             for region in regions:
@@ -53,78 +74,5 @@ class Processor(object):
                             
                             yield CrossProduct(sample, region, obs, syst, template,)
     
-class InputManager(object):
-    def __init__(self, xps, cfg):
-        read_methods = {
-                        'parquet': 'ak_parquet',
-                        'json': 'ak_json',
-                        'root': 'uproot',
-                        'h5': 'ak_h5'
-                        }
-        self.xps = xps
-        self.indirs = cfg["general"]["indir"]
-        self.from_pyth = cfg["general"]["frompythium"]
-        self.ext = cfg["general"]["informat"]
-        self.sample_sel = cfg["general"]["samplesel"]
-        self.reader = read_methods[self.ext]
-    
-    def required_variables(self):
-        req_vars: List[Observable] = []
-        xp_to_req = defaultdict(list)
-        for xp in self.xps:
-            sample, region, obs, syst, template = xp
-            required_variables = []
-            obs_vars, _region_sel_vars, sample_sel_vars, syst_vars = [],[],[],[]
-            
-            if obs.builder is None:   obs_vars = [obs]
-            else:
-                obs_vars = [Observable(reqvar, reqvar, obs.binning, obs.dataset) for reqvar in obs.builder.req_vars ]
 
-            region_sel_vars =  [ Observable(reqvar, reqvar, obs.binning, obs.dataset) for reqvar in region.sel.req_vars ] 
-            if self.sample_sel:
-                sample_sel_vars =  [ Observable(reqvar, reqvar, obs.binning, obs.dataset) for reqvar in sample.sel.req_vars ]
-            
-            if isinstance(systematic, WeightSyst):
-                template = getattr(systematic, template)
-                if isinstance(template, Functor):
-                    syst_vars =  [ Observable(reqvar, reqvar, obs.binning, obs.dataset) for reqvar in template.req_vars ] 
-                else:
-                    syst_vars = [Observable(template, template, obs.binning, obs.dataset)]  
-            
-            required_variables.extend([ obs_vars, region_sel_vars, sample_sel_vars, syst_vars])
-            
-            xp_to_req[xp] = required_variables
-            
-        return xp_to_req
-
-    def required_paths(self):
-        xp_to_paths = defaultdict(list)
-        for xp in self.xps:
-            sample, region, observable, systematic, template = xp
-            if self.from_pyth:  # Follow pythium naming scheme
-                sample_name = sample.name
-                obs_dataset = observable.dataset
-                paths = [f"{path}/{sample_name}_*_{obs_dataset}.{self.ext}" for path in self.indirs]
-                if template != 'nom':
-                    if isinstance(systematic, NTupSyst):
-                        sys_dirs =    getattr(systematic, "where")
-                        sys_samples = getattr(systematic, template)
-                        if sys_dirs == [None]:
-                            paths = [f"{indir}/{s_samp}_*_{obs_dataset}.{self.ext}" for indir in indirs for s_samp in sys_samples]
-                        else:
-                            paths = [f"{s_dir}/{s_samp}_*_{obs_dataset}.{self.ext}" for s_dir in sys_dirs for s_samp in sys_samples]
-                    elif (isinstance(systematic, TreeSyst)):
-                        syst_dataset = getattr(systematic, template) 
-                        paths = [f"{indir}/{sample_name}_*_{syst_dataset}.{self.ext}" for indir in self.indirs]
-
-                paths = list(set([p for path in paths for p in glob(path) if not os.path.isdir(p) ]))
-                xp_to_paths[xp] = paths
-               
-            
-            else:
-                ## TODO:: Assume user defined (somehow?) #Custom? some supported special types? decoreator for custom?
-                logger.warning("Only outputs from Pythium currently supported")
-                pass
-
-        return xp_to_paths
